:wrench: Add Nav bar in mission analysis
